// 04_Angular/Day7 - 15.02.2018/01_Book-store app + angular material/src/app/shared/models/book.model.ts
    export interface ImageLinks {
        smallThumbnail: string;
        thumbnail: string;
    }

    export interface VolumeInfo {
        title: string;
        authors?: string[];
        publisher?: string;
        publishedDate?: string;
        description?: string;
        pageCount?: number;
        imageLinks: ImageLinks;
        language?: string;
        subtitle?: string;
    }

    export interface ListPrice {
        amount?: number;
        currencyCode?: string;
    }

    export interface SaleInfo {
        listPrice?: ListPrice;
    }

    export interface Book {
        id?:string;
        volumeInfo: VolumeInfo;
        saleInfo?: SaleInfo;
    }

    export interface BookRootObject {
        items: Book[];
    }